выделение в отдельную компоненту HeaderNav
