+++ src/components/header/components/header-nav.tsx
@@ -1,0 +1,14 @@
+import HeaderUser from './header-user';
+import HeaderSign from './header-sign';
+
+function HeaderNav (): JSX.Element {
+  return (
+    <nav className="header__nav">
+      <ul className="header__nav-list">
+        <HeaderUser/>
+        <HeaderSign/>
+      </ul>
+    </nav>);
+}
+
+export default HeaderNav;
