+++ src/components/header/header.tsx
@@ -1,19 +1,8 @@
-import HeaderUser from './components/header-user';
-import HeaderSign from './components/header-sign';
 import HeaderLogo from './components/header-logo';
+import HeaderNav from './components/header-nav';
 
 type HeaderNavProps = {
   isHeaderNav?: boolean;
-}
-
-function HeaderNav (): JSX.Element {
-  return (
-    <nav className="header__nav">
-      <ul className="header__nav-list">
-        <HeaderUser/>
-        <HeaderSign/>
-      </ul>
-    </nav>);
 }
 
 function Header ({isHeaderNav = true}: HeaderNavProps):JSX.Element {
